+++ async_customerio/api.py
@@ -4,10 +4,17 @@ Implements the client that interacts with Customer.io"s App API using app keys.
 import base64
 import typing as t
 
+from typing_extensions import TypedDict
+
 from async_customerio.client_base import AsyncClientBase
 from async_customerio.errors import AsyncCustomerIOError
 from async_customerio.regions import Region, Regions
 from async_customerio.utils import join_url
+
+
+IdentifierID = TypedDict("IdentifierID", {"id": t.Union[str, int]})
+IdentifierEMAIL = TypedDict("IdentifierEMAIL", {"email": str})
+IdentifierCIOID = TypedDict("IdentifierCIOID", {"cio_id": t.Union[str, int]})
 
 
 class SendEmailRequest:
@@ -17,22 +24,22 @@ class SendEmailRequest:
         self,
         transactional_message_id: t.Union[str, int] = None,
         to: str = None,
-        identifiers=None,
+        identifiers: t.Union[IdentifierID, IdentifierEMAIL, IdentifierCIOID] = None,
         _from: str = None,
-        headers=None,
+        headers: t.Dict[str, str] = None,
         reply_to: str = None,
-        bcc=None,
+        bcc: str = None,
         subject: str = None,
-        preheader=None,
-        body=None,
+        preheader: str = None,
+        body: str = None,
         plaintext_body: str = None,
-        amp_body=None,
-        fake_bcc=None,
+        amp_body: str = None,
+        fake_bcc: str = None,
         disable_message_retention: bool = None,
         send_to_unsubscribed: bool = None,
         tracked: bool = None,
-        queue_draft=None,
-        message_data=None,
+        queue_draft: bool = None,
+        message_data: dict = None,
         attachments: t.Dict[str, str] = None,
     ):