+++ CHANGES.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## 0.3.0
+
+- ``AsyncClientBase`` has been adjusted to handle properly non-successful codes and connection retries.
+- Class ``SendEmailRequest`` has been updated with type annotations.
+
 ## 0.2.0
 
-  * `README.md` has been updated with examples of how to use.
+- `README.md` has been updated with examples of how to use.
 
 ## 0.1.0
 
-  * First release on PyPI.
+- First release on PyPI.